+++ database/factories/PostFactory.php
@@ -1,6 +1,8 @@
 <?php
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
+
+use App\Album;
 use App\Chest;
 use App\Link;
 use App\Post;
@@ -27,6 +29,10 @@ $factory->state(Post::class, 'chest', function (Faker $faker) {
     return ['postable_type' => Chest::class,];
 });
 
+$factory->state(Post::class, 'album', function (Faker $faker) {
+    return ['postable_type' => Album::class,];
+});
+
 $factory->state(Post::class, 'private', function (Faker $faker) {
     return [
         'is_private' => 1,
@@ -40,6 +46,8 @@ $factory->afterCreating(Post::class, function (Post $post, Faker $faker) {
         $post->postable_id = factory(Story::class)->create()->id;
     } else if ($post->postable_type == Chest::class) {
         $post->postable_id = factory(Chest::class)->create()->id;
+    } else if ($post->postable_type == Album::class) {
+        $post->postable_id = factory(Album::class)->create()->id;
     }
     $post->save();
 });