+++ resources/js/components/ChestCard.vue
@@ -19,24 +19,55 @@
             <span><i class="fas fa-lock pr-2" v-if="chest.is_private"></i>{{ chest.date_formated }}</span>
 
             <div class="dropdown">
-                <button class="btn btn-outline-dark btn-sm dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                <button class="btn btn-outline-dark btn-sm dropdown-toggle"
+                        type="button"
+                        id="dropdownMenuButton"
+                        data-toggle="dropdown"
+                        aria-haspopup="true"
+                        aria-expanded="false"
+                        dusk="chest-card-more"
+                >
                     {{ __('More') }}
                 </button>
 
                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
-                    <a class="dropdown-item" :href="chest.permalink"><i class="fas fa-link fa-fw mr-1"></i> {{ __('Permalink') }}</a>
-                    <button type="button" class="dropdown-item" @click="copyAll"><i class="fas fa-copy fa-fw mr-1"></i> {{ __('Copy all') }}</button>
+                    <a class="dropdown-item"
+                       :href="chest.permalink"
+                       dusk="chest-card-permalink"
+                    >
+                        <i class="fas fa-link fa-fw mr-1"></i> {{ __('Permalink') }}
+                    </a>
+                    <button type="button"
+                            class="dropdown-item"
+                            @click="copyAll"
+                            dusk="chest-card-copy"
+                    >
+                        <i class="fas fa-copy fa-fw mr-1"></i> {{ __('Copy all') }}
+                    </button>
 
                     <h6 class="dropdown-header" v-if="chest.editable">{{ __('Manage') }}</h6>
 
-                    <a class="dropdown-item" @click="$bus.$emit('share', chest)" v-if="chest.editable"><i class="fas fa-share-square fa-fw mr-1"></i> {{ __('Temp sharing') }}</a>
-                    <a class="dropdown-item" :href="chest.url_edit" v-if="chest.editable"><i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}</a>
+                    <a class="dropdown-item"
+                       @click="$bus.$emit('share', chest)"
+                       v-if="chest.editable"
+                       dusk="chest-card-temp-share"
+                    >
+                        <i class="fas fa-share-square fa-fw mr-1"></i> {{ __('Temp sharing') }}
+                    </a>
+                    <a class="dropdown-item"
+                       :href="chest.url_edit"
+                       v-if="chest.editable"
+                       dusk="chest-card-edit"
+                    >
+                        <i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}
+                    </a>
 
                     <confirm class="dropdown-item"
                              :text="`<i class='fas fa-trash-alt fa-fw mr-1'></i> ${__('Delete')}`"
                              :text-confirm="`<i class='fas fa-check fa-fw mr-1'></i> ${__('Confirm')}`"
                              @confirmed="remove"
                              v-if="chest.editable"
+                             dusk="chest-card-remove"
                     ></confirm>
                 </div>
             </div>