+++ resources/js/components/LinkCard.vue
@@ -20,25 +20,62 @@
             <span><i class="fas fa-lock pr-2" v-if="link.is_private"></i>{{ link.date_formated }}</span>
 
             <div class="dropdown">
-                <button class="btn btn-outline-dark btn-sm dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                <button class="btn btn-outline-dark btn-sm dropdown-toggle"
+                        type="button"
+                        id="dropdownMenuButton"
+                        data-toggle="dropdown"
+                        aria-haspopup="true"
+                        aria-expanded="false"
+                        dusk="link-card-more"
+                >
                     {{ __('More') }}
                 </button>
 
                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
-                    <a class="dropdown-item" :href="link.permalink"><i class="fas fa-link fa-fw mr-1"></i> {{ __('Permalink') }}</a>
-                    <a class="dropdown-item" :href="waybackUrl" target="_blank"><i class="fas fa-history fa-fw mr-1"></i> {{ __('View on archive.org') }}</a>
-                    <a class="dropdown-item" :href="link.url_download" v-if="link.url_download"><i class="fas fa-file-download fa-fw mr-1"></i> {{ __('Download') }}</a>
+                    <a class="dropdown-item"
+                       :href="link.permalink"
+                       dusk="link-card-permalink"
+                    >
+                        <i class="fas fa-link fa-fw mr-1"></i> {{ __('Permalink') }}
+                    </a>
+                    <a class="dropdown-item"
+                       :href="waybackUrl"
+                       target="_blank"
+                       dusk="link-card-archiveorg-link"
+                    >
+                        <i class="fas fa-history fa-fw mr-1"></i> {{ __('View on archive.org') }}
+                    </a>
+                    <a class="dropdown-item"
+                       :href="link.url_download"
+                       v-if="link.url_download"
+                       dusk="link-card-download-link"
+                    >
+                        <i class="fas fa-file-download fa-fw mr-1"></i> {{ __('Download') }}
+                    </a>
 
                     <h6 class="dropdown-header" v-if="link.editable">{{ __('Manage') }}</h6>
 
-                    <a class="dropdown-item" @click="$bus.$emit('share', link)" v-if="link.editable"><i class="fas fa-share-square fa-fw mr-1"></i> {{ __('Temp sharing') }}</a>
-                    <a class="dropdown-item" :href="link.url_edit" v-if="link.editable"><i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}</a>
+                    <a class="dropdown-item"
+                       @click="$bus.$emit('share', link)"
+                       v-if="link.editable"
+                       dusk="link-card-temp-share"
+                    >
+                        <i class="fas fa-share-square fa-fw mr-1"></i> {{ __('Temp sharing') }}
+                    </a>
+                    <a class="dropdown-item"
+                       :href="link.url_edit"
+                       v-if="link.editable"
+                       dusk="link-card-edit"
+                    >
+                        <i class="fas fa-pen-alt fa-fw mr-1"></i> {{ __('Edit') }}
+                    </a>
 
                     <confirm class="dropdown-item"
                              :text="`<i class='fas fa-trash-alt fa-fw mr-1'></i> ${__('Delete')}`"
                              :text-confirm="`<i class='fas fa-check fa-fw mr-1'></i> ${__('Confirm')}`"
                              @confirmed="remove"
                              v-if="link.editable"
+                             dusk="link-card-remove"
                     ></confirm>
                 </div>
             </div>